Moved sigmoid activation to it's own file, mod-ed the rest of the activation functions
