+++ src/activation/mod.rs
@@ -1,5 +1,9 @@
 use std::f64;
 
+pub mod sigmoid;
+
+pub use self::sigmoid::Sigmoid;
+
 /// Activation functions
 pub trait Activation {
     fn new() -> Self;
@@ -7,24 +11,6 @@ pub trait Activation {
     fn calc(&self, x: f64) -> f64;
     // Derivative
     fn derivative(&self, x: f64) -> f64;
-}
-
-pub struct Sigmoid;
-
-impl Activation for Sigmoid {
-    fn new() -> Sigmoid {
-        return Sigmoid;
-    }
-
-    /// Calculates the Sigmoid of input `x`
-    fn calc(&self, x: f64) -> f64 {
-        1f64 / (1f64 + (-x).exp())
-    }
-
-    /// Calculates the Derivative Sigmoid of input `x`
-    fn derivative(&self, x: f64) -> f64 {
-        x * (1f64 - x)
-    }
 }
 
 pub struct Identity;
@@ -111,23 +97,10 @@ impl Activation for RectifiedLinearUnit {
 #[cfg(test)]
 mod tests {
     use super::Activation;
-    use super::Sigmoid;
     use super::Identity;
     use super::HyperbolicTangent;
     use super::SoftPlus;
     use super::RectifiedLinearUnit;
-
-    #[test]
-    fn sigmoid_test() {
-        let activation = Sigmoid::new();
-        assert_approx_eq!(activation.calc(5f64), 0.9933071490f64);
-    }
-
-    #[test]
-    fn sigmoid_derivative_test() {
-        let activation = Sigmoid::new();
-        assert_approx_eq!(activation.derivative(5f64), -20f64);
-    }
 
     #[test]
     fn identity_test() {
